YER-65 [개발] 반환값에 page 정보가 필요한 경우 재가공하는 로직을 추가한다.

+++ src/main/java/com/freeboard02/api/PageDto.java
@@ -26,7 +26,7 @@ public class PageDto<T> {
     }
 
     public static <T> PageDto of(int totalDataSize, Pageable pageable, List<T> contents) {
-        int totalPages = (int) Math.ceil(((long) totalDataSize) / pageable.getPageSize());
+        int totalPages = (int) Math.ceil(((float) totalDataSize) / pageable.getPageSize());
         int nowPage = pageable.getPageNumber() + 1;
         int startPage = 1 * VIEWPAGESIZE * (nowPage / VIEWPAGESIZE);
         int endPage = startPage + VIEWPAGESIZE - 1;
